+++ CHANGELOG.md
@@ -24,6 +24,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Feature
+
+- **core:** create a functional guard `canMatchFeatureFlag`
+- **core:** add the configuration of the `canMatchFeatureFlag` behavior to the `NgxFlagrConfiguration`
+- **core:** update the usage of `provideNgxConfiguration` to include a default `routing` section
+
+### Others
+
+- **core:** move the definition of the `FeatureFlag` type into its own file
+- **core:** add a method to evaluate if a value is a `FeatureFlag`
+- **core:** add tests for the `provideNgxConfiguration` method
+
 ## [0.1.0] - (2023-04-16)
 
 ### Features